server/requirements.txt
asgiref==3.11.1
attrs==26.1.0
cffi==2.0.0
Django==6.0.4
djangorestframework==3.17.1
drf-spectacular==0.29.0
googleapis-common-protos==1.74.0
grpcio==1.80.0
gunicorn==25.3.0
importlib_metadata==8.7.1
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-exporter-prometheus==0.62b0
opentelemetry-instrumentation==0.62b0
opentelemetry-instrumentation-django==0.62b0
opentelemetry-instrumentation-logging==0.62b0
opentelemetry-instrumentation-wsgi==0.62b0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
opentelemetry-util-http==0.62b0
packaging==26.1
prometheus_client==0.25.0
protobuf==6.33.6
pycparser==3.0
pyroscope-io==1.0.5
python-json-logger==4.1.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
wrapt==2.1.2
zipp==3.23.1
